# Make slider path parsing more friendly with spans