--- a/osu.Game/Rulesets/Objects/PathControlPoint.cs
+++ b/osu.Game/Rulesets/Objects/PathControlPoint.cs
@@ -76,5 +76,9 @@ namespace osu.Game.Rulesets.Objects
         }
 
         public bool Equals(PathControlPoint other) => Position == other?.Position && Type == other.Type;
+
+        public override string ToString() => type is null
+            ? $"Position={Position}"
+            : $"Position={Position}, Type={type}";
     }
 }