In order to unblock #23282, this PR refactors slider parsing to not passing substrings around.

The majority of existing logic is preserved. The performance shouldn't change much with this PR, just ensuring there's no regression.

Before:

| Method                  | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 1.529 ms | 0.0083 ms | 0.0065 ms | 228.5156 | 187.5000 |   4.13 MB |

After:

 | Method                  | Mean     | Error     | StdDev    | Gen0     | Gen1     | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 1.540 ms | 0.0208 ms | 0.0195 ms | 228.5156 | 187.5000 |   4.12 MB |