@frenzibyte 2024-03-19
Requesting a second pair of eyes before proceeding with a merge, the PR is good but I'm not sure if there's something I should've checked against before approval (like triggering a diffcalc command to make sure nothing has changed with beatmap parsing or something).